before:

$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/tensorflow/tensorflow/tensorflow/go
BenchmarkNewTensor/[150528]-8                200           6792809 ns/op
PASS
ok      github.com/tensorflow/tensorflow/tensorflow/go  2.116s

after:

$ go test -bench=.
goos: linux
goarch: amd64
pkg: github.com/tensorflow/tensorflow/tensorflow/go
BenchmarkNewTensor/[150528]-8                500           3269740 ns/op
PASS
ok      github.com/tensorflow/tensorflow/tensorflow/go  2.021s