@yongtang 2017-11-09
It seems the `//tensorflow/go:test` test is failing as Uint32 and Uint64 has not been supported.

See Ln 74 of `tensorflow/go/tensor_test.go`

https://github.com/tensorflow/tensorflow/blob/5fb5d8207b576e57acd2fac00da429284eb26647/tensorflow/go/tensor_test.go#L74-L78

and compare Ln 334 and 316 of `tensorflow/go/tensor.go`:

https://github.com/tensorflow/tensorflow/blob/5fb5d8207b576e57acd2fac00da429284eb26647/tensorflow/go/tensor.go#L316

https://github.com/tensorflow/tensorflow/blob/5fb5d8207b576e57acd2fac00da429284eb26647/tensorflow/go/tensor.go#L334


Created a PR #14408 for the fix.